requirements.txt
attrs==21.4.0
aws-cdk-lib==2.9.0
cattrs==1.10.0
constructs==10.0.47
jsii==1.52.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.0.1
